+++ application/controllers/Fraksi.php
@@ -167,11 +167,18 @@ class Fraksi extends CI_Controller
         $row = $this->Fraksi_model->get_by_id($id);
 
         if ($row) {
-            $this->Fraksi_model->delete($id);
-            $this->session->set_flashdata('message', 'Delete Record Success');
-            redirect(site_url('fraksi'));
+            if ($this->session->userdata('id_group') == 1) {
+                $this->Fraksi_model->delete($id);
+                $this->session->set_flashdata('message', 'Delete Record Success');
+                redirect(site_url('fraksi'));
+            } else {
+
+                $this->Fraksi_model->update($id, ['isactive' => 0]);
+                $this->session->set_flashdata('message', 'Data Berhasil Dihapus');
+                redirect(site_url('fraksi'));
+            }
         } else {
-            $this->session->set_flashdata('message', 'Record Not Found');
+            $this->session->set_flashdata('message', 'Data Tidak Ditemukan');
             redirect(site_url('fraksi'));
         }
     }
